requirements.txt
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
dj-stripe==2.2.2
Django==2.2.3
django-allauth==0.41.0
django-appconf==1.0.3
django-better500s==0.2.1
django-braces==1.14.0
django-ckeditor==5.7.1
django-countries==6.1.3
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-image-cropping==1.2.0
django-js-asset==1.2.2
django-model-utils==4.0.0
django-money==1.1
django-paypal==1.0.0
django-ratings==0.3.7
django-star-ratings==0.9.1
easy-thumbnails==2.6
gunicorn==19.9.0
idna==2.8
jsonfield2==3.1.0
oauthlib==3.1.0
Pillow==6.1.0
psycopg2==2.8.3
py-moneyed==0.8.0
PyJWT==1.7.1
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.3.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
stripe==2.42.0
swapper==1.1.2.post1
urllib3==1.25.3
virtualenv==16.7.9
whitenoise==4.1.4
